+++ python/cusignal/filtering/__init__.py
@@ -19,10 +19,13 @@ from cusignal.filtering.resample import (
 )
 from cusignal.filtering.filtering import (
     wiener,
+    firfilter,
     lfiltic,
     sosfilt,
     hilbert,
     hilbert2,
     detrend,
+    channelize_poly,
+    channelize_poly_gpu,
     freq_shift,
 )